backend/requirements.txt
fastapi
uvicorn
kiteconnect
pandas
numpy
ta
python-dotenv
